inbound: Extract stacks into a separate files
This change creates new inbound::http and inbound::direct modules that
handle stack constructions.

No functional change.
